Pull inner ChatMessage class up to its own class
